Feat/auth (#9)
* Add user_settings.tsx file.

Will have to add a check for session id in index.tsx to check if the user is logged in. If there is a user logged in, signup and sign in pages should be replaced with a user settings button/link.

* Made basic UI template for User Settings page.

Completed the UI for side bar in the user settings. Will complete the main user settings UI next and also set up events based on button clicks in the side bar.

* Create UserSettingsContent and added it as a component that is returned in UserSettingsMain

UserSettingsContent contains the user settings content that should be displayed to the user depending on which option they choose on UserSettingsSideBar. I am also committing because Ed just merged to main and I have to commit the changes I have made.

* Organise user setting components into a file named "user_setting" under components directory.

* Made a dynamic navbar; changes depending on user's logged in status.

I installed react-firebase-hook library to help me get data on the current logged in user. Because
I could not get docRef.onSnapshot to work with typesaurus, I just used the firestore methods. I also created a
lib directory that contains the context.ts file for Context API for next.js and a hooks.tsx file which contains
the custom hook I made. I made it a custom hook so _app.tsx looks cleaner.