Feat/search page (#13)
* Add package 'Validator'

The package 'Validator' is installed. This allows for the validation of data, notably phone numbers, emails, URLs.

* Make changes to the firestoreTypes to reflect changes in Group structure

The group structure is changed, see commit for details.
The group id is not the same as the name, to avoid checking for duplicates later.
A new type 'Links' is added to cover the socials of each Group in a single data structure.

* Multiple components are added to support the creation of social links in the group page.

The UI for [groupName] is changed. Links to socials are now displayed.
Components are created to facilitate the re-use of the links across the site.

* Change firestoreTypes so Groups have meeting times and age ranges.

Update the Group type with the above properties.
Create Range and MeetingTime interfaces for these properties.
Create Day and Frequency enums to support the MeetingTime interface.

* Move component styling into a new file and add a style for grey text

Rename file from linkstyle to componentStyles.
Rename underline to underlineLinkHover for clarity.
Add style for grey text.

* Change LinkWithIcon and TextWithIcon to support new styling and add string conversion methods for the new interfaces

Update style references on LinkWithIcon and TextWithIcon and add title attribute support
Create StringConverters and populate with methods to convert the new data types to strings
Remove console.log from WhatsAppLink

* Add UI for AgeRange and MeetingTimes for Group

Add a component to display the Range of ages of a Group.
Add a component to display the MeetingTime of a Group, or a modal in the case of multiple MeetingTimes.
Update the UI of [groupName].

* Remove another errant console.log on WhatsAppLink

* Add a modal to show multiple meeting times

Add a modal on MeetingTimeDisplay
BUG: There is a non-breaking bug related to keys on list items
Add a function to stringConverters for time ranges
Change componentStyles to change the mouse pointer on links

* Update firestoreTypes for the GroupCategory

Change GroupTypes to GroupCategory and add appropriate categories
Change Links to SocialLinks for clarity
Add comments to Day and Frequency enums for clarity

* Create a ScoutIcon

Create a ScoutIcon as there is no built in icon
Create a media folder to place the ScoutIcon in

* Create various helper files

Create helperFunctions and add a title function
Create helperTypes and move the Host type into it
Create iconMappings to map types to icons
Map the Host and GroupCategory types
Update ValidatedLinkWithIcon with the above changes

* Add GroupCategory to UI

Create storageHelpers and use it to generate Google Storage URLs
Rename groupName to groupID as the name is NOT the ID
Create a GroupCategoryDisplay component
Add the GroupCategoryDisplay to the UI
Replace the placeholder image in the UI with the group avatar

* Add styled, but unfunctional, search box component

* Add helper for setting group avatar
Remove description from objectdefaults

* Limit description length on groupcard

* Add navbar to group page

* Add navbar to groupPage
Add theming to groupPage

Co-authored-by: Eliot <eliotpersonal77@gmail.com>